Use header line in CSV input
Note to reviewers: I had forgotten this part and had to fix it. I would
normally have squashed the two corresponding commits before submitting
them for review/testing/integration review, but have chosen not to do so
this time in order to allow you to see a part of the software
development livecycle.