Ignore invalid records
If value in the minimum column is larger than the value in the maximum
column, there is a mistake in programming (column selection) or in the
input data. Ignore the record in question.

Note to reviewers: if logging was implemented, a warning message would
have been appropriate. Adding a logger just for this case would probably
exceed the allocated time frame and desired code complexity and has
therefore been omitted.