Add football analysis
Second part of the challenge. Reuse the code which has already been
written for the first part (weather data analysis).

Note to reviewers: I am not much of a football guy. I have looked up
“goal difference” on Wikipedia, and chosen to implement it so that the
goal difference is the number of goals scored against the opponent minus
the number of goals allowed (and not the other way round). Hope that’s
what was intended. (I would usually have asked the customer about an
exact definition in case of doubt.)